♻️ Refactor: Rename grid utility tests and migrate position utility functions
