+++ packages/web/src/views/Calendar/components/Event/Grid/GridEvent/GridEvent.tsx
@@ -17,13 +17,13 @@ import {
   FlexDirections,
   FlexWrap,
 } from "@web/components/Flex/styled";
-import { getPosition } from "@web/views/Calendar/hooks/event/getPosition";
 import { getLineClamp } from "@web/common/utils/grid.util";
 import { getTimesLabel } from "@web/common/utils/web.date.util";
 import { ZIndex } from "@web/common/constants/web.constants";
 import { Text } from "@web/components/Text";
 
 import { StyledEvent, StyledEventScaler, StyledEventTitle } from "../../styled";
+import { getPosition } from "@web/common/utils/position.util";
 
 interface Props {
   event: Schema_GridEvent;
