+++ packages/web/src/views/Calendar/components/Grid/AllDayRow/AllDayEvent.tsx
@@ -4,9 +4,9 @@ import { Priorities } from "@core/constants/core.constants";
 import { Flex } from "@web/components/Flex";
 import { Schema_GridEvent } from "@web/common/types/web.event.types";
 import { isOptimisticEvent } from "@web/common/utils/event.util";
+import { getPosition } from "@web/common/utils/position.util";
 import { AlignItems, FlexDirections } from "@web/components/Flex/styled";
 import { SpaceCharacter } from "@web/components/SpaceCharacter";
-import { getPosition } from "@web/views/Calendar/hooks/event/getPosition";
 import { Measurements_Grid } from "@web/views/Calendar/hooks/grid/useGridLayout";
 import { WeekProps } from "@web/views/Calendar/hooks/useWeek";
 import { Text } from "@web/components/Text";
